refactor: apply form 스키마 및 타입 구조 개선 (#46)
